Endrer arbeidssituasjon til aktivitet
- Sånn at det passer kontrakter bedre
