+++ src/models/aktivitet/aktivitet.ts
@@ -2,12 +2,14 @@ import { ITekstListeFelt, ITekstFelt } from '../søknadsfelter';
 import { IUnderUtdanning } from './utdanning';
 import { IArbeidsgiver } from './arbeidsgiver';
 import { IArbeidssøker } from './arbeidssøker';
+import { IFirma } from './firma';
 
-export interface IArbeidssituasjon {
-  situasjon: ITekstListeFelt;
+export interface IAktivitet {
+  hvaErDinArbeidssituasjon: ITekstListeFelt;
   etablererEgenVirksomhet?: ITekstFelt;
   arbeidsforhold?: IArbeidsgiver[];
   arbeidssøker?: IArbeidssøker;
+  firma?: IFirma;
   underUtdanning?: IUnderUtdanning;
 }
