+++ src/models/aktivitet/firma.ts
@@ -1,0 +1,15 @@
+import { IDatoFelt, ITekstFelt } from '../søknadsfelter';
+
+export interface IFirma {
+  navn?: ITekstFelt;
+  organisasjonsnummer?: ITekstFelt;
+  etableringsdato?: IDatoFelt;
+  arbeidsmengde?: ITekstFelt;
+  arbeidsuke?: ITekstFelt;
+}
+
+export enum EFirma {
+  navn = 'navn',
+  organisasjonsnummer = 'organisasjonsnummer',
+  arbeidsmengde = 'arbeidsmengde',
+}
