+++ src/models/søknad.ts
@@ -1,4 +1,4 @@
-import { IArbeidssituasjon } from './arbeidssituasjon/arbeidssituasjon';
+import { IAktivitet } from './aktivitet/aktivitet';
 import { IPerson } from './person';
 import { IVedlegg } from './vedlegg';
 import { IMedlemskap, ISivilstatus } from './omDeg';
@@ -12,7 +12,7 @@ export interface ISøknad {
   sivilstatus: ISivilstatus;
   medlemskap: IMedlemskap;
   bosituasjon: IBosituasjon;
-  arbeidssituasjon: IArbeidssituasjon;
+  aktivitet: IAktivitet;
   vedleggsliste: IVedlegg[];
 }
