+++ src/søknad/Søknadsdialog.tsx
@@ -1,12 +1,12 @@
 import React, { FC, useEffect } from 'react';
 import { Route, Switch } from 'react-router';
-import OmDeg from './steg/1-omdeg/OmDeg';
-import SendSøknad from './SendSøknad';
-import Forside from './forside/Forside';
+import Aktivitet from './steg/5-aktivitet/Aktivitet';
+import BarnaDine from './steg/3-barnadine/BarnaDine';
 import BarnasBosted from './steg/4-barnasbosted/BarnasBosted';
 import Bosituasjon from './steg/2-bosituasjon/Bosituasjon';
-import Arbeidssituasjon from './steg/5-arbeidssituasjon/Arbeidssituasjon';
-import BarnaDine from './steg/3-barnadine/BarnaDine';
+import Forside from './forside/Forside';
+import OmDeg from './steg/1-omdeg/OmDeg';
+import SendSøknad from './SendSøknad';
 
 const Søknadsdialog: FC = () => {
   useEffect(() => {
@@ -20,7 +20,7 @@ const Søknadsdialog: FC = () => {
         <Route path={'/bosituasjon'} component={Bosituasjon} />
         <Route path={'/barn'} component={BarnaDine} />
         <Route path={'/barnas-bosted'} component={BarnasBosted} />
-        <Route path={'/arbeidsituasjon'} component={Arbeidssituasjon} />
+        <Route path={'/aktivitet'} component={Aktivitet} />
         <Route path={'/send-soknad'} component={SendSøknad} />
         <Route path={'/'} component={Forside} />
       </Switch>
