+++ src/søknad/steg/5-aktivitet/Aktivitet.tsx
@@ -6,11 +6,11 @@ import CheckboxSpørsmål from '../../../components/spørsmål/CheckboxSpørsmå
 import HjemmeMedBarnUnderEttÅr from './HjemmeMedBarnUnderEttÅr';
 import EtablererEgenVirksomhet from './EtablererEgenVirksomhet';
 import OmArbeidsforholdetDitt from './arbeidsforhold/OmArbeidsforholdetDitt';
-import { hvaErDinArbeidssituasjon } from './ArbeidssituasjonConfig';
+import { hvaErDinArbeidssituasjonSpm } from './AktivitetConfig';
 import {
   EArbeidssituasjon,
-  IArbeidssituasjon,
-} from '../../../models/arbeidssituasjon/arbeidssituasjon';
+  IAktivitet,
+} from '../../../models/aktivitet/aktivitet';
 import UnderUtdanning from './underUtdanning/UnderUtdanning';
 import Arbeidssøker from './arbeidssøker/Arbeidssøker';
 import SeksjonGruppe from '../../../components/gruppe/SeksjonGruppe';
@@ -19,20 +19,20 @@ import { hentTekst } from '../../../utils/søknad';
 import { nyttTekstListeFelt } from '../../../utils/søknadsfelter';
 import OmFirmaetDitt from './OmFirmaetDitt';
 
-const Arbeidssituasjon: React.FC = () => {
+const Aktivitet: React.FC = () => {
   const intl = useIntl();
   const { søknad, settSøknad } = useSøknadContext();
-  const [arbeidssituasjon, settArbeidssituasjon] = useState<IArbeidssituasjon>({
-    situasjon: nyttTekstListeFelt,
+  const [arbeidssituasjon, settArbeidssituasjon] = useState<IAktivitet>({
+    hvaErDinArbeidssituasjon: nyttTekstListeFelt,
   });
-  const { situasjon } = arbeidssituasjon;
+  const { hvaErDinArbeidssituasjon } = arbeidssituasjon;
 
   useEffect(() => {
-    settSøknad({ ...søknad, arbeidssituasjon: arbeidssituasjon });
+    settSøknad({ ...søknad, aktivitet: arbeidssituasjon });
     // eslint-disable-next-line
   }, [arbeidssituasjon]);
 
-  const oppdaterArbeidssituasjon = (nyArbeidssituasjon: IArbeidssituasjon) => {
+  const oppdaterArbeidssituasjon = (nyArbeidssituasjon: IAktivitet) => {
     settArbeidssituasjon({ ...arbeidssituasjon, ...nyArbeidssituasjon });
   };
 
@@ -49,7 +49,7 @@ const Arbeidssituasjon: React.FC = () => {
   const erAktivitetHuketAv = (aktivitet: EArbeidssituasjon): boolean => {
     const tekstid: string = 'arbeidssituasjon.svar.' + aktivitet;
     const svarTekst: string = intl.formatMessage({ id: tekstid });
-    return situasjon.verdi.some((svarHuketAvISøknad: string) => {
+    return hvaErDinArbeidssituasjon.verdi.some((svarHuketAvISøknad: string) => {
       return svarHuketAvISøknad === svarTekst;
     });
   };
@@ -78,9 +78,9 @@ const Arbeidssituasjon: React.FC = () => {
     <Side tittel={intl.formatMessage({ id: 'stegtittel.arbeidssituasjon' })}>
       <SeksjonGruppe>
         <CheckboxSpørsmål
-          spørsmål={hvaErDinArbeidssituasjon}
+          spørsmål={hvaErDinArbeidssituasjonSpm}
           settValgteSvar={settArbeidssituasjonFelt}
-          valgteSvar={situasjon?.verdi}
+          valgteSvar={hvaErDinArbeidssituasjon?.verdi}
         />
       </SeksjonGruppe>
 
@@ -124,4 +124,4 @@ const Arbeidssituasjon: React.FC = () => {
   );
 };
 
-export default Arbeidssituasjon;
+export default Aktivitet;
