+++ src/søknad/steg/5-aktivitet/arbeidsforhold/ArbeidsgiverConfig.ts
@@ -2,7 +2,7 @@ import { ISpørsmål, JaNeiSvar } from '../../../../models/spørsmal';
 import {
   EArbeidsgiver,
   EStilling,
-} from '../../../../models/arbeidssituasjon/arbeidsgiver';
+} from '../../../../models/aktivitet/arbeidsgiver';
 
 export const hvaSlagsStilling: ISpørsmål = {
   søknadid: EArbeidsgiver.fastStilling,
