+++ src/søknad/steg/5-aktivitet/arbeidsforhold/HarSøkerSluttdato.tsx
@@ -12,7 +12,7 @@ import FeltGruppe from '../../../../components/gruppe/FeltGruppe';
 import {
   EArbeidsgiver,
   IArbeidsgiver,
-} from '../../../../models/arbeidssituasjon/arbeidsgiver';
+} from '../../../../models/aktivitet/arbeidsgiver';
 
 interface Props {
   arbeidsgiver: IArbeidsgiver;
