+++ src/søknad/steg/5-aktivitet/underUtdanning/ErUtdanningenOffentligEllerPrivat.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import MultiSvarSpørsmål from '../../../../components/spørsmål/MultiSvarSpørsmål';
-import { heltidEllerDeltidSpm } from './UtdanningConfig';
+import { privatEllerOffentligSpm } from './UtdanningConfig';
 import KomponentGruppe from '../../../../components/gruppe/KomponentGruppe';
-import {
-  EUtdanning,
-  IUnderUtdanning,
-} from '../../../../models/arbeidssituasjon/utdanning';
+import { IUnderUtdanning } from '../../../../models/aktivitet/utdanning';
 import { ISpørsmål } from '../../../../models/spørsmal';
 import { hentTekst } from '../../../../utils/søknad';
 import { useIntl } from 'react-intl';
@@ -13,26 +10,13 @@ interface Props {
   utdanning: IUnderUtdanning;
   settUtdanning: (utdanning: IUnderUtdanning) => void;
 }
-const ErUtdanningenPåHeltidEllerDeltid: React.FC<Props> = ({
+const ErUtdanningenOffentligEllerPrivat: React.FC<Props> = ({
   utdanning,
   settUtdanning,
 }) => {
   const intl = useIntl();
 
   const settMultiSpørsmål = (spørsmål: ISpørsmål, svar: string) => {
-    const søkerVilStudereHeltid = spørsmål.svaralternativer.find(
-      (svarsalternativ) =>
-        hentTekst(svarsalternativ.svar_tekstid, intl) === svar
-    );
-    if (
-      (spørsmål.søknadid === EUtdanning.heltidEllerDeltid &&
-        søkerVilStudereHeltid &&
-        utdanning.målMedUtdanning) ||
-      utdanning.arbeidsmengde
-    ) {
-      delete utdanning.arbeidsmengde;
-      delete utdanning.målMedUtdanning;
-    }
     settUtdanning({
       ...utdanning,
       [spørsmål.søknadid]: {
@@ -44,13 +28,13 @@ const ErUtdanningenPåHeltidEllerDeltid: React.FC<Props> = ({
   return (
     <KomponentGruppe>
       <MultiSvarSpørsmål
-        spørsmål={heltidEllerDeltidSpm}
+        spørsmål={privatEllerOffentligSpm}
         settSpørsmålOgSvar={settMultiSpørsmål}
-        valgtSvar={utdanning.heltidEllerDeltid?.verdi}
+        valgtSvar={utdanning.offentligEllerPrivat?.verdi}
         toKorteSvar={true}
       />
     </KomponentGruppe>
   );
 };
 
-export default ErUtdanningenPåHeltidEllerDeltid;
+export default ErUtdanningenOffentligEllerPrivat;
